+++ test/test_dalli.rb
@@ -26,6 +26,17 @@ class TestDalli < Test::Unit::TestCase
     assert_equal s2, s3
   end
 
+  should "have the continuum sorted by value" do
+    servers = [stub(:hostname => "localhost", :port => "11211", :weight => 1),
+               stub(:hostname => "localhost", :port => "9500", :weight => 1)]
+    ring = Dalli::Ring.new(servers, {})
+    previous_value = 0
+    ring.continuum.each do |entry|
+      assert entry.value > previous_value
+      previous_value = entry.value
+    end
+  end
+
   context 'using a live server' do
 
     should "support get/set" do